fix data type

+++ data/src/main/java/pe/devpicon/android/codelab/pomodoro/data/remote/response/RandomActivityResponse.kt
@@ -1,11 +1,11 @@
 package pe.devpicon.android.codelab.pomodoro.data.remote.response
 
 data class RandomActivityResponse(
-    val accessibility: Int,
+    val accessibility: Double,
     val activity: String,
     val key: String,
     val link: String,
     val participants: Int,
-    val price: Int,
+    val price: Double,
     val type: String
 )
